Moved dividing the message into fours into shift module
