+++ enigma/lib/cipher.rb
@@ -10,13 +10,13 @@ class Cipher
     @date = date
   end
 
-  def div_message_into_fours
-    four_characters = []
-    @message.chars.each_slice(4) do |chars|
-      four_characters.push(chars)
-    end
-    four_characters
-  end
+  # def div_message_into_fours
+  #   four_characters = []
+  #   @message.chars.each_slice(4) do |chars|
+  #     four_characters.push(chars)
+  #   end
+  #   four_characters
+  # end
 
   def encrypt
     message_split = div_message_into_fours
